backend/requirements.txt
fastapi
uvicorn
python-dotenv
groq
supabase
google-auth
google-auth-oauthlib
google-api-python-client
requests